fix(pkg/kubelego) do not match ingress class with provider name
When using a custom LEGO_DEFAULT_INGRESS_CLASS and
LEGO_SUPPORTED_INGRESS_CLASS ingress resources where not being applied
due to the matching on the ingress class name and the kube-lego provider
name.

By applying the ingress resource on every provider according to the
LEGO_SUPPORTED_INGRESS_CLASS it is possible to use custom ingress class
names on kube-lego.